# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Willard AI — NAS Media & Data Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /media/cleanup:
    post: {<<: *runtimeJsonMutationOperation, operationId: createMediaCleanup}
...